refactor(configuration): :recycle: add dto in use case users
