+++ src/application/use_cases/register_user.rs
@@ -1,6 +1,6 @@
 use crate::domain::repositories::user_repository::UserRepository;
 use crate::domain::services::user_service::UserService;
-use crate::presentation::handlers::user_handler::NewUser;
+use crate::presentation::dtos::users_dto::NewUser;
 
 pub struct RegisterUserUseCase<T: UserRepository> {
     user_service: UserService<T>
